api/types: move HijackedResponse into client
